Extract useReviewRows into its own module to break circular import
